Separate files for *_P functions
Don't link the whole .o when only one function is used
