+++ newlib/libc/sys/xtensa/pgmspace/sprintf_P.c
@@ -1,0 +1,13 @@
+/* Backwards compatibility, both format and arguments are allowed to be PSTR / PROGMEM */
+
+#include <stdarg.h>
+#include <stdio.h>
+
+int sprintf_P(char* str, const char* formatP, ...) {
+    int ret;
+    va_list arglist;
+    va_start(arglist, formatP);
+    ret = vsprintf(str, formatP, arglist);
+    va_end(arglist);
+    return ret;
+}
