+++ newlib/libc/sys/xtensa/pgmspace/strncat_P.c
@@ -1,0 +1,55 @@
+/*
+ Modified from original source by Earle F. Philhower, III
+ Original authorship:
+ Copyright (c) 2015 Michael C. Miller.  All right reserved.
+
+ This library is free software; you can redistribute it and/or
+ modify it under the terms of the GNU Lesser General Public
+ License as published by the Free Software Foundation; either
+ version 2.1 of the License, or (at your option) any later version.
+
+ This library is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ Lesser General Public License for more details.
+
+ You should have received a copy of the GNU Lesser General Public
+ License along with this library; if not, write to the Free Software
+ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+ */
+
+// TODO - Optimize these routines to use 32-bit accesses whenever possible
+
+#define __need_size_t
+#include <stddef.h>
+
+#include <string.h>
+#include <sys/pgmspace.h>
+
+char* strncat_P(char* dest, const char* src, size_t size)
+{
+    char* write = dest;
+
+    while (*write != '\0')
+    {
+        write++;
+    }
+
+    const char* read = src;
+    char ch = '.';
+
+    while (size > 0 && ch != '\0')
+    {
+        ch = pgm_read_byte(read++);
+        *write++ = ch;
+
+        size--;
+    }
+
+    if (ch != '\0')
+    {
+        *write = '\0';
+    }
+
+    return dest;
+}
