+++ newlib/libc/sys/xtensa/pgmspace/vsnprintf_P.c
@@ -1,0 +1,9 @@
+/* Backwards compatibility, both format and arguments are allowed to be PSTR / PROGMEM */
+
+#include <stdio.h>
+
+int vsnprintf_P(char* str, size_t strSize, const char* formatP, __gnuc_va_list ap) {
+    int ret;
+    ret = vsnprintf(str, strSize, formatP, ap);
+    return ret;
+}
